feat: set up rating component (#68)
This pull request introduces a new `Rating` component to the
`MediaInformation` section and includes some responsive design
adjustments. Here are the most important changes:

### New Feature: Rating Component

*
[`src/frontend/src/components/Rating/Rating.jsx`](diffhunk://#diff-952c6b33e1548b29ee68b2999a7039d6c02e057835bf1fdf45df32d408a43e9aR1-R19):
Added a new `Rating` component that displays a list of ratings with
their respective names and scores.
*
[`src/frontend/src/components/Rating/Rating.css`](diffhunk://#diff-97feee953ee38e3bc246dca8608edce6ed20cd95a4f880190ce933b86a5cc8e4R1-R22):
Created a new CSS file for the `Rating` component to style the ratings
and ensure responsiveness.
*
[`src/frontend/src/components/index.js`](diffhunk://#diff-812f2eaaa6f2e0da446fafe1709eb999dcbaf4fd5bb150c7f14933f2822a8a56R20):
Exported the new `Rating` component for use in other parts of the
application.

### Integration of Rating Component

*
[`src/frontend/src/components/MediaInformation/MediaInformation.jsx`](diffhunk://#diff-1c6dba6e61786d471185066fa8b0804d3e551ce373e9ebaddaa916c075dbae12L11-R11):
Imported the `Rating` component and integrated it into the
`MediaInformation` component, including updating the props to accept
`ratings`.
[[1]](diffhunk://#diff-1c6dba6e61786d471185066fa8b0804d3e551ce373e9ebaddaa916c075dbae12L11-R11)
[[2]](diffhunk://#diff-1c6dba6e61786d471185066fa8b0804d3e551ce373e9ebaddaa916c075dbae12R23)
[[3]](diffhunk://#diff-1c6dba6e61786d471185066fa8b0804d3e551ce373e9ebaddaa916c075dbae12L115-R130)
*
[`src/frontend/src/pages/MediaDetailPage.jsx`](diffhunk://#diff-9067da5199ca855ad6e34c7c96e7704c01a34040a5826da405606ca2edcb83ecR16-R21):
Added a sample `ratings` array to the dummy data for the
`MediaDetailPage` to demonstrate the new `Rating` component.

### Responsive Design Adjustments

*
[`src/frontend/src/components/MediaInformation/MediaInformation.css`](diffhunk://#diff-bfc6fb4fb15c098fadf25b28a6af486f6d443147a64b1cf939af7c8958d0e89dR25-R34):
Added media queries to adjust the layout of the crew information section
for smaller screens.