feat: implementation of pagination component done. (#58)
This pull request introduces a new `Pagination` component to the
frontend. The component is designed to handle pagination for lists of
items, making it easier to navigate through large datasets.

The most important changes include:

Pagination component:

*
[`src/frontend/src/components/Pagination/Pagination.jsx`](diffhunk://#diff-91b141d548ef43421e370c6d79e71519634d498810a79ffe97811a095b7d7039R1-R71):
Added a new `Pagination` component using `react-bootstrap` to handle
pagination logic and rendering. The component calculates the total
number of pages, handles page changes, and generates pagination items
including ellipses for skipped pages.

Component export:

*
[`src/frontend/src/components/index.js`](diffhunk://#diff-812f2eaaa6f2e0da446fafe1709eb999dcbaf4fd5bb150c7f14933f2822a8a56R20):
Added an export for the new `Pagination` component to make it available
for use throughout the application.