feat: auth session validation (#78)
This pull request includes several changes to improve the handling of
user authentication and simplify path references in the frontend
project. The most important changes include updating the `AuthProvider`
component to handle session storage changes and adding path aliases in
the `jsconfig.json` file.

Improvements to user authentication:

*
[`src/frontend/src/contexts/AuthContext.jsx`](diffhunk://#diff-e293315d942711a42e6d81dc270374876410251fb776ea749df677e3de314d7eL26-R48):
Updated the `useEffect` hook to check for both token and stored user,
set `isAuthenticated` to false if either is missing, and added an event
listener to handle session storage changes.
*
[`src/frontend/src/contexts/AuthContext.jsx`](diffhunk://#diff-e293315d942711a42e6d81dc270374876410251fb776ea749df677e3de314d7eR69):
Modified the `logout` function to set `isAuthenticated` to false when a
user logs out.

Simplification of path references:

*
[`src/frontend/jsconfig.json`](diffhunk://#diff-6fbeaa032e9c2d52effad0fec65089fb7346972042141728476d10f2bb8fea0aR1-R7):
Added path aliases to simplify module imports using `@/` as a shorthand
for `./src/`.