feat: set up logic for following a person (#81)
This pull request introduces the functionality to follow a person on the
`PersonDetailPage` in the frontend. The most important changes include
adding a follow button, implementing the follow action, and handling the
follow request in the service layer.

### Changes to `PersonDetailPage`:

* Added `Button` import from `react-bootstrap` and `createFollow` import
from `userService` to `PersonDetailPage.jsx`.
* Implemented `handleFollow` function to handle the follow action,
including error handling and toast messages.
* Added a follow button to the `PersonDetailPage` UI, which triggers the
`handleFollow` function.

### Changes to service layer:

* Added `createFollow` function to `userService.js` to handle the follow
request to the backend.