feat: person card + grid and set up af pages (#83)
This pull request includes several changes to improve the structure and
functionality of the frontend components, particularly those related to
media and person entities. The most important changes include the
addition of new components, refactoring existing components to use a
more generic hook, and updating routes and styles.

### Component Additions and Updates:
*
[`src/frontend/src/components/PersonsCarousel/PersonsCarousel.jsx`](diffhunk://#diff-f124bc23b307bb6de3f29438dcd1a02ed871c1c0743fde0a59d89fa672aeb3bbR1-R38):
Added a new `PersonsCarousel` component to display a carousel of
persons.
*
[`src/frontend/src/components/PersonsGrid/PersonsGrid.jsx`](diffhunk://#diff-358e0ee73b6b9099027e612bf257e1e04c483af465810499d3dd2dadce5d808fR1-R74):
Added a new `PersonsGrid` component to display a grid of persons.
*
[`src/frontend/src/components/PersonCard/PersonCard.jsx`](diffhunk://#diff-cbf9b6845f8d8582081ae680fc285eaa8832f0827a1403f0f3f453c23e540d33L17-R17):
Updated the `PersonCard` component to navigate to the new `persons`
route.

### Hook Refactoring:
*
[`src/frontend/src/hooks/index.js`](diffhunk://#diff-d83bde231cf2c823d085b178febbb7cb65877477acbcfd46622b21d4d3e4bf56L3-R4):
Replaced `useMediaPerRow` with a more generic `useItemsPerRow` hook.
*
[`src/frontend/src/hooks/useItemsPerRow.js`](diffhunk://#diff-693aaa5a3dc7c1e5e6c27cd6e89385bfd8e1edb3c9d52b512f1203a87ec0021eR1-R35):
Added a new `useItemsPerRow` hook to determine the number of items per
row based on screen width.
*
[`src/frontend/src/hooks/useMediaPerRow.js`](diffhunk://#diff-2573c67909b8dc52fb0e9607d6e49cb9c7c7e42d4581312e84ad5f77736088edL1-L35):
Removed the `useMediaPerRow` hook as it is now replaced by
`useItemsPerRow`.

### Route and Style Updates:
*
[`src/frontend/src/App.jsx`](diffhunk://#diff-02cfd7a23e4f0e1e03bb8ccc3d578d916c7331eb7b5847bf71d18fa02f4489a9L32-R40):
Updated routes to include nested routes for media and persons, and added
a new `PersonOverviewPage` route.
*
[`src/frontend/src/global.css`](diffhunk://#diff-5b91e2deb2467ebec6641c03e5f02d80ce1a8bcaafa7cc6bb6b938b0c5240416R6-R28):
Moved carousel control and button styles from individual component CSS
files to the global CSS file.

### Page Enhancements:
*
[`src/frontend/src/pages/PersonDetailPage.jsx`](diffhunk://#diff-410656ca5388cae2df048377b54b05faa0a9fe86341734904e2831769706285cL1-R102):
Enhanced the `PersonDetailPage` to include a carousel of co-actors and
handle loading states.
*
[`src/frontend/src/pages/PersonOverviewPage.jsx`](diffhunk://#diff-1c2170ac31f8d87243e15d106f3bf8f4d76ae5d7e35f5a40404189a3036e6560R1-R34):
Added a new `PersonOverviewPage` to display a list of persons using the
new `PersonsCarousel` and `PersonsGrid` components.