feat: Index.js (#93)
This pull request to the `src/frontend/src/components` directory
includes changes that reorganize import statements for better
readability and maintainability. The most important changes include
moving imports to group them logically, ensuring third-party imports are
separated from local imports, and consolidating imports from the same
module.

Reorganization of import statements:

*
[`src/frontend/src/components/BookmarkList/BookmarkList.jsx`](diffhunk://#diff-4e9400a52fdedad8c3b85878bea8813f5df88185cc71ebb6f8ddbd2b8c92f1c9R1-L4):
Moved `Link` import to the top of the file to group it with other
third-party imports.
*
[`src/frontend/src/components/CompletedList/CompletedList.jsx`](diffhunk://#diff-75c22b5ac12eb67d58ea444c573e029ae54ac143859b22da6f85d2e0a48b300aR1-L4):
Moved `Link` import to the top of the file to group it with other
third-party imports.
*
[`src/frontend/src/components/DynamicBreadcrumb/DynamicBreadcrumb.jsx`](diffhunk://#diff-4951685e76d6b9138a19ebdb9638edc99ee3b0fe1834edfdbc4000affdae84faL1-R2):
Reordered imports to place the `Breadcrumb` import after the `Link`
import.
*
[`src/frontend/src/components/FilterMediaComponent/FilterMediaComponent.jsx`](diffhunk://#diff-b9334d5015e73f217521eccb718c55abbddb49131215747e36c4e1c17f3504f3L1-R2):
Reordered imports to place the `Form`, `Row`, `Col`, and `Button`
imports after the `useState` import.
*
[`src/frontend/src/components/FollowingList/FollowingList.jsx`](diffhunk://#diff-9195c10ad88b718849c3070389f9570ef414038657bdf709d334e7ce2c21a176R1-L4):
Moved `Link` import to the top of the file to group it with other
third-party imports.