fix: fix console errors and depend on the is authenticated to fetch data (#118)
This pull request includes updates to the `SearchForm` component to
improve its functionality and ensure proper handling of user
authentication. The most important changes include adding a new hook,
modifying the search history fetching logic, and updating the error
message for better clarity.

Improvements to `SearchForm` component:

*
[`src/frontend/src/components/SearchForm/SearchForm.jsx`](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L2-R12):
Added the `useCallback` hook to optimize the `fetchSearchHistory`
function and prevent unnecessary re-renders.
[[1]](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L2-R12)
[[2]](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L28-R35)
[[3]](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L38-R49)
*
[`src/frontend/src/components/SearchForm/SearchForm.jsx`](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L2-R12):
Integrated `useAuth` hook to check if the user is authenticated before
fetching search history.
[[1]](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L2-R12)
[[2]](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L28-R35)
*
[`src/frontend/src/components/SearchForm/SearchForm.jsx`](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L38-R49):
Updated the error message in the `fetchSearchHistory` function to be
more concise and user-friendly.
*
[`src/frontend/src/components/SearchForm/SearchForm.jsx`](diffhunk://#diff-d63c9b4d81c977c527ab14269171efd40dff36598801da6596c94eb85680aa76L56-R65):
Modified the `useEffect` dependency array to include
`fetchSearchHistory` to ensure it is called when dependencies change.