fix: login styling (#129)
This pull request includes changes to the layout and styling of the
`LoginForm` and `NavBar` components in the frontend. The most important
changes include adjustments to the width of the login form and dropdown
menu, as well as the addition of a responsive design for smaller
screens.

Styling updates:

*
[`src/frontend/src/components/LoginForm/LoginForm.css`](diffhunk://#diff-6107dfe4fdcda393a8af14a926f0ece518481636976cc57196d0caa73aed7ab3L2):
Removed the `min-width` property from the `.login-form` class, ensuring
the login form takes up the full available width.
*
[`src/frontend/src/components/NavBar/NavBar.css`](diffhunk://#diff-8be15b8475d9d1341111739f871b109a9bd309425b96e2a86c160ed0c65c3489R9-R25):
Added styles for the `.dropdown-menu.show` and `.login-container`
classes to set their widths and make them responsive to screen size
changes.

Component updates:

*
[`src/frontend/src/components/NavBar/Navbar.jsx`](diffhunk://#diff-fc8b5b77a7589f2a9c4fbba5ce959590059ce850594f91edc990ca1664689741L75-R75):
Applied the `login-container` class to the `div` containing the
`LoginForm` component to align with the new styling rules.